Fix general parser issues with unknown tags
This commit fixes issues with the parser when it encounters an unknown
tag in the GEDCOM file. Previously, the parser would note the error and
continue to the next line. This was causing deeper data to be associated
to the last level, instead of ignoring the whole block.

The parser will now skip forward to the next chunk of data, essentially
ignoring the entire block of unknown data.

Also added in parsing INDI\[EVEN]\CHAN for good measure.

Fixes #12